Build and Deploy a Crowdfund Application
Utilizing the programming language solidity and the remix ide website. A straightforward develop and deploy crowdfund application is made in this case. ERC20 tokens were also utilised in the development of this application. When we perform a search for IERC20 “Ethereum Request for Comments” on the internet, we may locate this file.www.patika.dev